update build script to copy local.settings.php into place if it’s present in the PROJECT/scripts/settings directory - also updating code that previously copied site.settings.php to prevent error if site.settings.php isn’t present

+++ scripts/build.sh
@@ -155,8 +155,18 @@ else
 fi
 
 SETTINGS_SITE="$DESTINATION/profiles/$PROJECT/scripts/settings/site.settings.php"
-cp $SETTINGS_SITE $DESTINATION/sites/default/site.settings.php
-echo "Copied site.settings.php into place."
+if [ -f $SETTINGS_SITE ]; then
+  cp $SETTINGS_SITE $DESTINATION/sites/default/site.settings.php
+  printf "Copied site.settings.php into place.\n"
+fi
+
+SETTINGS_LOCAL="$DESTINATION/profiles/$PROJECT/scripts/settings/local.settings.php"
+if [ -f $SETTINGS_LOCAL ]; then
+  # replace PROJECT with actual project name in local.settings.php
+  sed -i '' "s/PROJECT/$PROJECT/g" $DESTINATION/profiles/$PROJECT/scripts/settings/local.settings.php
+  cp $SETTINGS_LOCAL $DESTINATION/sites/default/local.settings.php
+  printf "Copied local.settings.php into place.\n"
+fi
 
 # uncomment RewriteBase in project's .htaccess file
 # necessary for Drupal sites running on a machine configured using ThinkShout standards
